Restructured again to make imports work
