+++ src/planet-model-2/layers/core.py
@@ -1,3 +1,5 @@
+from layers.layer import layer
+
 class core:
 
     def __init__(self):
